django/requirements.txt
Django>=1.11,<1.12
djangorestframework>=3.6,<3.7
django-codemirror-widget>=0.4,<0.5

flake8>=3.4,<3.5

pytest>=3.2,<3.3
pytest-django>=3.1,<3.2
pytest-flake8>=0.8,<0.9

pytest-cov>=2.5,<2.6
celery>=4.1,<4.2
docker>=2.5,<2.6
lti>=0.8.4,<0.8.5
pyyaml>=3.12,<3.13
